requirements.txt
meraki
argparse
logging
azure-functions